chore: inherit the npm version from alpine-node
reinstalling npm 3 on docker results in a`npm ERR! EXDEV: cross-device link not permitted`. this is a known issue that does not seem to get addressed anytime soon, see https://github.com/npm/npm/issues/9863

node 4 shipped with npm 2 by default, so we had to reinstall npm3. as node 6 ships with npm 3 and the alpine image contains the latest npm version, we can get rid of the update step